Moved testcase.config and webdriver.config back to the config subpackage
After looking through the remaining configs, it seems to make more sense to put them back in this location, especially since most configs are used by the testcase package and some configs don't fit neatly into any of the project subpackages (e.g. site and test).
