+++ webdriver_test_tools/config/browserstack.py
@@ -1,7 +1,6 @@
 """
-This module imports the :class:`BrowserStackConfig <webdriver_test_tools.testcase.config.BrowserStackConfig>`
-class for compatibility with legacy test projects. See the documentation for
-:mod:`webdriver_test_tools.testcase.config` for details.
+This module imports the :class:`BrowserStackConfig <webdriver_test_tools.config.browser.BrowserStackConfig>`
+class for consistency with test project configs. See the documentation for
+:mod:`webdriver_test_tools.config.browser`.
 """
-from webdriver_test_tools.testcase.config import BrowserStackConfig
-
+from .browser import BrowserStackConfig
